requirements.txt
aiohttp
aioboto3
ollama
torch
transformers
nano_vectordb
bilibili_api_dev
fastapi==0.115.6
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.28
langchain_openai==0.2.2
langgraph==0.2.60
langserve==0.3.1
lightrag_hku==1.0.8
networkx==3.2.1
numpy==1.26.4
pydantic==2.10.4
python-dotenv==1.0.1
pyvis==0.3.2
streamlit==1.39.0
typing_extensions==4.12.2
uvicorn==0.34.0
